+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Adds missing properties for `log` in `cds.env`
 
+### Removed
+- [breaking] Removed type `TypedRequest<T>` and replaced it with just `Request<T>`
+
+
 ### Fixed
 
 - Use `Required` instead of `DeepRequired` in projection function to avoid complexity errors from TypeScript